simple-msa/order-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.2
pydantic==2.6.1
requests==2.31.0 